Move CommonFormEx and IFormView from Common/Controls to Common/SystemUtil (#2642)
* Moved CommonFormEx and IFormView from Common/Controls to Common/SystemUtil

* Fix Code Inspection test

* Remove redundant using directives.

---------
