+++ pwiz_tools/Shared/Common/Common.csproj
@@ -113,13 +113,9 @@
     <Compile Include="Controls\CommonDataGridView.cs">
       <SubType>Component</SubType>
     </Compile>
-    <Compile Include="Controls\CommonFormEx.cs">
-      <SubType>Form</SubType>
-    </Compile>
     <Compile Include="Controls\DoubleBufferedListView.cs">
       <SubType>Component</SubType>
     </Compile>
-    <Compile Include="Controls\IFormView.cs" />
     <Compile Include="Controls\ColumnLimitExceededColumn.cs" />
     <Compile Include="DataAnalysis\alglib\alglibinternal.cs" />
     <Compile Include="DataAnalysis\alglib\alglibmisc.cs" />
