+++ pwiz_tools/Shared/Common/DataBinding/Controls/Editor/DocumentationViewer.cs
@@ -18,7 +18,7 @@
  */
 using System;
 using System.Windows.Forms;
-using pwiz.Common.Controls;
+using pwiz.Common.SystemUtil;
 
 namespace pwiz.Common.DataBinding.Controls.Editor
 {
