+++ pwiz_tools/Shared/Common/DataBinding/Controls/Editor/FindColumnDlg.cs
@@ -19,7 +19,7 @@
 using System;
 using System.Collections.Generic;
 using System.Windows.Forms;
-using pwiz.Common.Controls;
+using pwiz.Common.SystemUtil;
 
 namespace pwiz.Common.DataBinding.Controls.Editor
 {
