+++ pwiz_tools/Shared/Common/DataBinding/Controls/Editor/ViewEditor.cs
@@ -24,7 +24,6 @@ using System.Drawing;
 using System.Linq;
 using System.Windows.Forms;
 using pwiz.Common.Collections;
-using pwiz.Common.Controls;
 using pwiz.Common.DataBinding.Documentation;
 using pwiz.Common.Properties;
 using pwiz.Common.SystemUtil;
