+++ pwiz_tools/Shared/Common/DataBinding/Controls/ManageLayoutsForm.cs
@@ -19,10 +19,10 @@
 using System.Drawing;
 using System.Linq;
 using System.Windows.Forms;
-using pwiz.Common.Controls;
 using pwiz.Common.DataBinding.Controls.Editor;
 using pwiz.Common.DataBinding.Layout;
 using pwiz.Common.Properties;
+using pwiz.Common.SystemUtil;
 
 namespace pwiz.Common.DataBinding.Controls
 {
