+++ pwiz_tools/Shared/Common/DataBinding/Controls/PivotEditor.cs
@@ -23,9 +23,9 @@ using System.Drawing;
 using System.Linq;
 using System.Windows.Forms;
 using pwiz.Common.Collections;
-using pwiz.Common.Controls;
 using pwiz.Common.DataBinding.Layout;
 using pwiz.Common.Properties;
+using pwiz.Common.SystemUtil;
 
 namespace pwiz.Common.DataBinding.Controls
 {
