+++ pwiz_tools/Skyline/Controls/Graphs/AreaGraphController.cs
@@ -21,7 +21,7 @@ using System;
 using System.Linq;
 using System.Windows.Forms;
 using pwiz.Common.Collections;
-using pwiz.Common.Controls;
+using pwiz.Common.SystemUtil;
 using pwiz.Skyline.Model;
 using pwiz.Skyline.Model.Results;
 using pwiz.Skyline.Properties;
