+++ pwiz_tools/Skyline/Controls/Graphs/DetectionsGraphController.cs
@@ -23,7 +23,6 @@ using System.Linq;
 using System.Reflection;
 using System.Windows.Forms;
 using pwiz.Common.Collections;
-using pwiz.Common.Controls;
 using pwiz.Common.SystemUtil;
 using pwiz.Skyline.Model;
 using pwiz.Skyline.Properties;
