+++ pwiz_tools/Skyline/Program.cs
@@ -31,7 +31,6 @@ using System.Threading;
 using System.Windows.Forms;
 using Microsoft.Win32;
 using pwiz.Common.Collections;
-using pwiz.Common.Controls;
 using pwiz.Common.SystemUtil;
 using pwiz.Skyline.Alerts;
 using pwiz.Skyline.Controls;
