+++ pwiz_tools/Skyline/SettingsUI/EditPeakScoringModelDlg.cs
@@ -23,7 +23,6 @@ using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Windows.Forms;
-using pwiz.Common.Controls;
 using ZedGraph;
 using pwiz.Common.DataBinding;
 using pwiz.Common.SystemUtil;
