+++ pwiz_tools/Skyline/ToolsUI/ToolOptionsUI.cs
@@ -24,7 +24,6 @@ using System.Globalization;
 using System.Linq;
 using System.Windows.Forms;
 using Grpc.Core;
-using pwiz.Common.Controls;
 using pwiz.Common.SystemUtil;
 using pwiz.Skyline.Alerts;
 using pwiz.Skyline.Controls.Graphs;
